Fix OpenBSD compilation errors
Those changes fix the following compilation errors on OpenBSD:

src/tofu.c:128:28: error: format specifies type 'long' but the argument has type
      'time_t' (aka 'long long') [-Werror,-Wformat]
                        "SHA-512", fingerprint, expires);

src/gmnlm.c:341:31: error: missing sentinel in function call
      [-Werror,-Wsentinel]
                execlp("sh", "sh", "-c", cmd);
                                            ^
                                            , NULL

+++ src/tofu.c
@@ -124,8 +124,8 @@ callback:
 		struct tm expires_tm;
 		ASN1_TIME_to_tm(notAfter, &expires_tm);
 		time_t expires = mktime(&expires_tm);
-		fprintf(f, "%s %s %s %ld\n", servername,
-			"SHA-512", fingerprint, expires);
+		fprintf(f, "%s %s %s %jd\n", servername,
+			"SHA-512", fingerprint, (intmax_t)expires);
 		fclose(f);
 
 		host = calloc(1, sizeof(struct known_host));
